Semi-automatic detection and enabling of TLS support.
If make detects the required OpenSSL files, it enable TLS support
in the build. This is keyed off the values in OPENSSLINC and
OPENSSLLIB. They default to the locations used by MacPorts. If
you use a different package manager, or build OpenSSL yourself,
you might need to change those two macros.